requirements.txt
cloudpickle==3.1.2
contourpy==1.3.2
cycler==0.12.1
Farama-Notifications==0.0.4
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
gym==0.25.1
gym-notices==0.1.0
gym-super-mario-bros==7.4.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
nes_py==8.2.1
networkx==3.4.2
numpy==1.24.4
packaging==25.0
pillow==12.0.0
pyglet==1.5.21
pyparsing==3.2.5
python-dateutil==2.9.0.post0
six==1.17.0
sympy==1.14.0
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
typing_extensions==4.15.0
